[charts] Extract plot logic into separate files for reuse
