+++ packages/x-charts/src/ScatterChart/Scatter.tsx
@@ -3,12 +3,7 @@ import * as React from 'react';
 import PropTypes from 'prop-types';
 import useSlotProps from '@mui/utils/useSlotProps';
 import { ScatterMarkerSlotProps, ScatterMarkerSlots } from './ScatterMarker.types';
-import {
-  DefaultizedScatterSeriesType,
-  ScatterItemIdentifier,
-  ScatterValueType,
-} from '../models/seriesType/scatter';
-import { getValueToPositionMapper } from '../hooks/useScale';
+import { DefaultizedScatterSeriesType, ScatterItemIdentifier } from '../models/seriesType/scatter';
 import { useInteractionAllItemProps } from '../hooks/useInteractionItemProps';
 import { useStore } from '../internals/store/useStore';
 import { useSelector } from '../internals/store/useSelector';
@@ -18,11 +13,11 @@ import {
   selectorChartsVoronoiIsVoronoiEnabled,
   UseChartVoronoiSignature,
 } from '../internals/plugins/featurePlugins/useChartVoronoi';
-import { useChartContext } from '../context/ChartProvider';
 import { ScatterMarker } from './ScatterMarker';
-import { SeriesId } from '../models/seriesType/common';
 import { ColorGetter } from '../internals/plugins/models/seriesConfig';
 import { ScatterClasses, useUtilityClasses } from './scatterClasses';
+import { useScatterPlotData } from './useScatterPlotData';
+import { useChartContext } from '../context/ChartProvider';
 
 export interface ScatterProps {
   series: DefaultizedScatterSeriesType;
@@ -78,62 +73,9 @@ function Scatter(props: ScatterProps) {
   const skipInteractionHandlers = isVoronoiEnabled || series.disableHover;
   const { isFaded, isHighlighted } = useItemHighlightedGetter();
 
-  const cleanData = React.useMemo(() => {
-    const getXPosition = getValueToPositionMapper(xScale);
-    const getYPosition = getValueToPositionMapper(yScale);
+  const scatterPlotData = useScatterPlotData(series, xScale, yScale, instance.isPointInside);
 
-    const temp: (ScatterValueType & {
-      dataIndex: number;
-      color: string;
-      isHighlighted: boolean;
-      isFaded: boolean;
-      interactionProps?: ReturnType<typeof useInteractionAllItemProps>[0];
-      seriesId: SeriesId;
-      type: 'scatter';
-    })[] = [];
-
-    for (let i = 0; i < series.data.length; i += 1) {
-      const scatterPoint = series.data[i];
-
-      const x = getXPosition(scatterPoint.x);
-      const y = getYPosition(scatterPoint.y);
-
-      const isInRange = instance.isPointInside(x, y);
-
-      if (isInRange) {
-        const currentItem = {
-          seriesId: series.id,
-          dataIndex: i,
-        };
-        const isItemHighlighted = isHighlighted(currentItem);
-        temp.push({
-          x,
-          y,
-          isHighlighted: isItemHighlighted,
-          isFaded: !isItemHighlighted && isFaded(currentItem),
-          id: scatterPoint.id,
-          seriesId: series.id,
-          type: 'scatter',
-          dataIndex: i,
-          color: colorGetter ? colorGetter(i) : color,
-        });
-      }
-    }
-
-    return temp;
-  }, [
-    xScale,
-    yScale,
-    series.data,
-    series.id,
-    isHighlighted,
-    isFaded,
-    colorGetter,
-    color,
-    instance,
-  ]);
-
-  const interactionItemProps = useInteractionAllItemProps(cleanData, skipInteractionHandlers);
+  const interactionItemProps = useInteractionAllItemProps(scatterPlotData, skipInteractionHandlers);
 
   const Marker = slots?.marker ?? ScatterMarker;
   const { ownerState, ...markerProps } = useSlotProps({
@@ -150,30 +92,35 @@ function Scatter(props: ScatterProps) {
 
   return (
     <g data-series={series.id} className={classes.root}>
-      {cleanData.map((dataPoint, i) => (
-        <Marker
-          key={dataPoint.id ?? dataPoint.dataIndex}
-          dataIndex={dataPoint.dataIndex}
-          color={dataPoint.color}
-          isHighlighted={dataPoint.isHighlighted}
-          isFaded={dataPoint.isFaded}
-          x={dataPoint.x}
-          y={dataPoint.y}
-          onClick={
-            onItemClick &&
-            ((event) =>
-              onItemClick(event, {
-                type: 'scatter',
-                seriesId: series.id,
-                dataIndex: dataPoint.dataIndex,
-              }))
-          }
-          data-highlighted={dataPoint.isHighlighted || undefined}
-          data-faded={dataPoint.isFaded || undefined}
-          {...interactionItemProps[i]}
-          {...markerProps}
-        />
-      ))}
+      {scatterPlotData.map((dataPoint, i) => {
+        const isItemHighlighted = isHighlighted(dataPoint);
+        const isItemFaded = !isItemHighlighted && isFaded(dataPoint);
+
+        return (
+          <Marker
+            key={dataPoint.id ?? dataPoint.dataIndex}
+            dataIndex={dataPoint.dataIndex}
+            color={colorGetter ? colorGetter(i) : color}
+            isHighlighted={isItemHighlighted}
+            isFaded={isItemFaded}
+            x={dataPoint.x}
+            y={dataPoint.y}
+            onClick={
+              onItemClick &&
+              ((event) =>
+                onItemClick(event, {
+                  type: 'scatter',
+                  seriesId: series.id,
+                  dataIndex: dataPoint.dataIndex,
+                }))
+            }
+            data-highlighted={isItemHighlighted}
+            data-faded={isItemFaded}
+            {...interactionItemProps[i]}
+            {...markerProps}
+          />
+        );
+      })}
     </g>
   );
 }
